requirements.txt
streamlit
google-generativeai
python-dotenv


#this is the change i made in requirement.txt